Payments based on Strategy pattern implemented

+++ src/main/java/com/ecommerce/year2_sem2_project/Entity/Order.java
@@ -31,13 +31,13 @@ public class Order {
     private List<OrderedItem> orderedItems;
 
     @Column(nullable = false)
-    private long totalPrice;
+    private double totalPrice;
 
     public Order() {
         this.orderDate = new Date();
     }
 
-    public Order(String customerName, String email, String address, Date orderDate, List<OrderedItem> orderedItems, long totalPrice) {
+    public Order(String customerName, String email, String address, Date orderDate, List<OrderedItem> orderedItems, double totalPrice) {
         this.customerName = customerName;
         this.email = email;
         this.address = address;
@@ -98,7 +98,7 @@ public class Order {
         return totalPrice;
     }
 
-    public void setTotalPrice(long totalPrice) {
+    public void setTotalPrice(double totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
